+++ ProcessFileEdit.config.php
@@ -6,6 +6,51 @@
  */
 class ProcessFileEditConfig extends ModuleConfig {
 
+
+
+	/**
+	 * Recursive function to extend starting directory selection to include template subdirs
+	 *
+	 * @param string $path starting directory
+	 * @param array $options
+	 * @return array $options
+	 *
+	 */
+	protected static function scanForSubdirs($path, &$options) {
+		// Ignore ./, ../ and .<folder>/ paths...
+		$files = array_diff(@scandir($path), array('.', '..'));
+		foreach ($files as $f) {
+			if (is_dir("$path$f") && (0 !== strpos($f, '.'))) {
+				$options["$path$f"] = "$path$f";
+				self::scanForSubdirs("$path$f" . DIRECTORY_SEPARATOR, $options);
+			}
+		}
+	}
+
+	/**
+	 * Create a list of directories from which to get files.
+	 *
+	 * @param string $paths
+	 * @return array $options
+	 *
+	 */
+	protected static function getDirPathOptions($paths) {
+		// predefined paths
+		$options = array(
+			$paths->root => $paths->root,
+			$paths->site => $paths->site,
+			$paths->siteModules => $paths->siteModules,
+			$paths->templates => $paths->templates,
+		);
+
+		// add directories under /site/templates
+		if (function_exists('scandir')) {
+			self::scanForSubdirs($paths->templates, $options);
+		}
+
+		return $options;
+	}
+
 	public function __construct() {
 
 		$this->add(array(
@@ -17,12 +62,7 @@ class ProcessFileEditConfig extends ModuleConfig {
 				'description' => $this->_('Path to the directory from which to get files.'),
 				'columnWidth' => 50,
 				'required'    => true,
-				'options'     => array(
-					$this->config->paths->root => $this->config->paths->root,
-					$this->config->paths->site => $this->config->paths->site,
-					$this->config->paths->templates => $this->config->paths->templates,
-					$this->config->paths->siteModules => $this->config->paths->siteModules,
-				),
+				'options'     => self::getDirPathOptions($this->config->paths),
 				'value'       => $this->config->paths->site,
 			),
 
@@ -30,16 +70,35 @@ class ProcessFileEditConfig extends ModuleConfig {
 				'name'        => 'encoding',
 				'type'        => 'select',
 				'label'       => $this->_('File encoding'),
-				'description' => $this->_('In case file names are garbled, try different encoding.'),
-				'columnWidth' => 50,
+				'description' => $this->_('Encoding used when saving file name.'),
+				//'notes' => $this->_('In case file names are garbled, try different encoding.'),
+				'columnWidth' => 25,
 				'required'    => true,
 				'options'     => array(
-					'auto'         => 'Auto detect',
-					'Windows-1250' => 'Windows-1250',
-					'Windows-1252' => 'Windows-1252',
-					'ISO-8859-2'   => 'ISO-8859-2',
-					'urldecode'    => 'PHP\'s urldecode',
-					'none'         => 'none',
+					'auto'          => 'Auto detect',
+					'Windows-1250'  => 'Windows-1250',
+					'Windows-1252'  => 'Windows-1252',
+					'ISO-8859-2'    => 'ISO-8859-2',
+					'urldecode'     => 'PHP\'s urldecode',
+					'none'          => 'none',
+				),
+				'value'       => 'auto',
+			),
+
+			array(
+				'name'        => 'lineEndings',
+				'type'        => 'select',
+				'label'       => $this->_('Line Endings'),
+				'description' => $this->_('Type of line ending to use when saving.'),
+				//'notes'       => $this->_('Default is Auto.'),
+				'columnWidth' => 25,
+				'required'    => true,
+				'options'     => array(
+					'auto'          => 'Auto',
+					'win'           => 'Windows (\r\n)',
+					'mac'           => 'Mac (\r)',
+					'nix'           => 'Linux (\n)',
+					'none'          => 'none',
 				),
 				'value'       => 'auto',
 			),
@@ -63,8 +122,8 @@ class ProcessFileEditConfig extends ModuleConfig {
 				'columnWidth' => 50,
 				'required'    => true,
 				'options'     => array(
-					'0' 				=> $this->_('Include files with named extensions'),
-					'1' 				=> $this->_('Exclude files with named extensions'),
+					'0'             => $this->_('Include files with named extensions'),
+					'1'             => $this->_('Exclude files with named extensions'),
 				),
 				'value'       => '0',
 			),